split TemplateConfig, add readTemplateFile
