+++ src/TemplateConfig.ts
@@ -1,0 +1,10 @@
+export interface CopyListItem {
+	src: string;
+	dst?: string;
+}
+
+export interface TemplateConfig {
+	files?: CopyListItem[];
+	gameJson?: string;
+	guideMessage?: string;
+}
